Float Cast on EntityRendererAlt

+++ src/main/java/noppes/mpm/client/EntityRendererAlt.java
@@ -25,7 +25,7 @@ public class EntityRendererAlt extends EntityRenderer{
     	ModelData data = PlayerDataController.instance.getPlayerData(player);
 		player.yOffset -= data.offsetY();
 		if(data.animation == EnumAnimation.SITTING)
-			player.yOffset += 0.5f - data.getLegsY() * 0.8;
+			player.yOffset += (float) (0.5f - data.getLegsY() * 0.8);
 		if(data.isSleeping() || data.animation == EnumAnimation.CRAWLING)
 			player.yOffset = 2.8f - (player.isSneaking()?0.125f:0);
 		if(player.yOffset < 1.4f && isBlocked(player))
@@ -51,7 +51,7 @@ public class EntityRendererAlt extends EntityRenderer{
 
 		float offset = -data.offsetY();
 		if(data.animation == EnumAnimation.SITTING){
-			offset += 0.5f - data.getLegsY() * 0.8;
+			offset += (float) (0.5f - data.getLegsY() * 0.8);
 		}
 		if(data.isSleeping() || data.animation == EnumAnimation.CRAWLING)
 			offset = 1.18f;
